Move ZotCourse helpers to zotcourse.ts
